Extract shared CLI helpers to break import cycle for session.py
The session subcommand was importing _epilog/_resolve_db_path/
_resolve_settings from datasight.cli, which in turn imports the session
module inside _register_commands(). The runtime tolerates this via the
lazy registration call but CodeQL flags the static cycle and any new
subcommand inherits it.

Move those three helpers into a dedicated datasight.cli_helpers module
that imports nothing from datasight.cli, and have cli.py re-export them
so the existing 'from datasight.cli import _epilog' calls in other
subcommand modules continue to work unchanged. session.py now imports
straight from cli_helpers, breaking the back-edge for this file.
